Fixed bug where Header was not rerendering on Login
Added "setUser" function to App wrapper and passed it
as props to the header and the login function, that way
when the user logs in, this function is called from the app
component rather than only the signIn page. This causes the
header and login to rerender, showing us our "logged in" HTML
and not this when user is not authenticated.