Security update: Log-in Token creation: Front end
In order to abstract the log in credentials away
from the stored data, every time a user logs in
they recieve a unique token which is stored locally
and used to validate queries made while logged in.

This token is destroyed upon the logging out. Lastly,
this feature is not fully functional until the token is
deleted upon logging out.